Split out a base listing view from generic IndexView
An attempt to clean up the issue encountered in #10452 where IndexView is making assumptions about the nature of the items being listed (e.g. their permission checks) that don't apply to all listings.
